+++ @utahdts/utah-design-system/react/components/forms/ComboBox/internal/CombBoxListBox.jsx
@@ -1,5 +1,5 @@
 import { useCallback, useEffect, useRef } from 'react';
-import { useFloating, autoUpdate, offset, shift, flip } from '@floating-ui/react-dom';
+import { useFloating, autoUpdate, offset as floatingOffset, shift, flip } from '@floating-ui/react-dom';
 import { useAriaMessaging } from '../../../../contexts/UtahDesignSystemContext/hooks/useAriaMessaging';
 import { popupPlacement } from '../../../../enums/popupPlacement';
 import { useDebounceFunc } from '../../../../hooks/useDebounceFunc';
@@ -48,7 +48,7 @@ export function CombBoxListBox({
       floating: ulRef.current,
     },
     middleware: [
-      offset({mainAxis: 4, crossAxis: 0, alignmentAxis: 0}),
+      floatingOffset({mainAxis: 4, crossAxis: 0, alignmentAxis: 0}),
       flip(),
       shift(),
     ],