+++ @utahdts/utah-design-system/react/components/navigation/items/MenuItemFlyout.jsx
@@ -1,6 +1,6 @@
 import { popupFocusHandler } from '@utahdts/utah-design-system-header';
 import { useEffect, useRef } from 'react';
-import { autoUpdate, flip, offset, shift, useFloating } from '@floating-ui/react-dom';
+import { autoUpdate, flip, offset as floatingOffset, shift, useFloating } from '@floating-ui/react-dom';
 import { useImmer } from 'use-immer';
 import { ICON_BUTTON_APPEARANCE } from '../../../enums/buttonEnums';
 import { menuTypes } from '../../../enums/menuTypes';
@@ -41,7 +41,7 @@ export function MenuItemFlyout({
     open: isChildrenOpen,
     placement: 'right-start',
     middleware: [
-      offset(10),
+      floatingOffset(10),
       flip(),
       shift(),
     ],