+++ utah-design-system-website/src/react/components/websiteContent/library/components/popups/popups/PopupsDocumentation.jsx
@@ -608,7 +608,7 @@ export function PopupsDocumentation() {
                       </TableCell>
                     </TableRow>
                     <TableRow>
-                      <TableCell><code className="primary-color">position</code></TableCell>
+                      <TableCell><code className="primary-color">offset</code></TableCell>
                       <TableCell>
                         <code>number</code>
                         <span> | </span>